==> 4_data_analysis/0_preparing_the_dataset.ipynb <==
# Notebook 1, Preparing the Dataset for Time Series Modeling

This notebook prepares the rainfall and temperature dataset for machine learning modeling.
The dataset is originally in wide format, where each row represents one region and one year containing monthly values.
We reshape the dataset into long format, create lag features for rainfall and temperature, and export the cleaned dataset for modeling.

We model separate forecasts for each region, so structuring the time series correctly is essential.

### Step 1: Load the dataset

import pandas as pd


df = pd.read_csv("../1_datasets/Final_dataset/final_merged_dataset.csv")
df.head()



### Step 2: Reshape Dataset to Long Format

Your current columns are like:

JAN_RAIN, FEB_RAIN, ...
JAN_TEMP, FEB_TEMP, ...

We need:

Region, Year, Month, Rainfall, Temperature

This is required to create time series lags.

# Identify month names in order
months = ["JAN","FEB","MAR","APR","MAY","JUN","JUL","AUG","SEP","OCT","NOV","DEC"]

# Create Rainfall long format
rain_long = df.melt(
    id_vars=["REGION","YEAR"],
    value_vars=[m + "_RAIN" for m in months],
    var_name="Month_Rain",
    value_name="Rainfall"
)

# Create Temperature long format
temp_long = df.melt(
    id_vars=["REGION","YEAR"],
    value_vars=[m + "_TEMP" for m in months],
    var_name="Month_Temp",
    value_name="Temperature"
)

# Extract month from column names
rain_long["Month"] = rain_long["Month_Rain"].str.split("_").str[0]
temp_long["Month"] = temp_long["Month_Temp"].str.split("_").str[0]

# Merge on region, year and month
df_long = pd.merge(
    rain_long[["REGION","YEAR","Month","Rainfall"]],
    temp_long[["REGION","YEAR","Month","Temperature"]],
    on=["REGION","YEAR","Month"]
)

df_long.head()



### Step 3: Convert Month to Numerical Order and Sort

month_order = {
    "JAN":1,"FEB":2,"MAR":3,"APR":4,"MAY":5,"JUN":6,
    "JUL":7,"AUG":8,"SEP":9,"OCT":10,"NOV":11,"DEC":12
}

df_long["Month_Num"] = df_long["Month"].map(month_order)

df_long = df_long.sort_values(["REGION","YEAR","Month_Num"]).reset_index(drop=True)
df_long.head(14)


### Step 4: Create a Time Index

Machine learning models like having a continuous time column.

df_long["Time"] = df_long.groupby("REGION").cumcount()
df_long.head()


### Step 5: Create Lag Features

We generate:

lag 1, lag 2, lag 3, lag 12

for rainfall and temperature.

lags = [1,2,3,12]

df_lags = df_long.copy()

for lag in lags:
    df_lags[f"Rain_lag_{lag}"] = df_lags.groupby("REGION")["Rainfall"].shift(lag)
    df_lags[f"Temp_lag_{lag}"] = df_lags.groupby("REGION")["Temperature"].shift(lag)

df_lags.head(15)


### Step 6: We can now drop misssing lag rows

df_lags = df_lags.dropna().reset_index(drop=True)
df_lags.head()


### Step 7: Save the Processed Dataset

df_lags.to_csv("../4_data_analysis/model_datasets/model_ready_dataset.csv", index=False)
df_lags.head()
